+++ types/protocol.d.ts
@@ -10512,6 +10512,17 @@ export namespace Protocol {
             nodeId: DOM.NodeId;
         }
 
+        export interface FlexNodeHighlightConfig {
+            /**
+             * A descriptor for the highlight appearance of flex containers.
+             */
+            flexContainerHighlightConfig: FlexContainerHighlightConfig;
+            /**
+             * Identifier of the node to highlight.
+             */
+            nodeId: DOM.NodeId;
+        }
+
         /**
          * Configuration for dual screen hinge
          */
@@ -10734,6 +10745,13 @@ export namespace Protocol {
              * An array of node identifiers and descriptors for the highlight appearance.
              */
             gridNodeHighlightConfigs: GridNodeHighlightConfig[];
+        }
+
+        export interface SetShowFlexOverlaysRequest {
+            /**
+             * An array of node identifiers and descriptors for the highlight appearance.
+             */
+            flexNodeHighlightConfigs: FlexNodeHighlightConfig[];
         }
 
         export interface SetShowPaintRectsRequest {